healthcare_rag/requirements.txt
annotated-doc
annotated-types
anyio
asyncpg
certifi
click
colorama
distro
faiss-cpu
fastapi
greenlet
h11
httpcore
httptools
httpx
idna
jiter
numpy
openai
packaging
pydantic
pydantic-settings
pydantic_core
pypdf
python-dotenv
python-multipart
PyYAML
redis
sniffio
SQLAlchemy
starlette
tqdm
typing-inspection
typing_extensions
uvicorn
watchfiles
websockets
google-generativeai
redisvl
